:truck: Rename and move theme-component
