+++ @navikt/core/react/src/theme/index.ts
@@ -1,0 +1,2 @@
+"use client";
+export { Theme } from "./Theme";
