+++ @navikt/core/react/src/timeline/period/ClickablePeriod.tsx
@@ -15,7 +15,7 @@ import {
 } from "@floating-ui/react";
 import cl from "clsx";
 import React, { useRef, useState } from "react";
-import { UNSAFE_useAkselTheme } from "../../provider";
+import { useThemeInternal } from "../../theme/Theme";
 import { useMergeRefs } from "../../util/hooks/useMergeRefs";
 import { useI18n } from "../../util/i18n/i18n.context";
 import { usePeriodContext } from "../hooks/usePeriodContext";
@@ -56,7 +56,7 @@ const ClickablePeriod = React.memo(
     const arrowRef = useRef<HTMLDivElement | null>(null);
     const translate = useI18n("Timeline");
 
-    const themeContext = UNSAFE_useAkselTheme(false);
+    const themeContext = useThemeInternal();
     const showArrow = !themeContext;
 
     const {
